requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
llama-cpp-python==0.2.56
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
sentence-transformers==2.2.2
numpy==1.24.3
